requirements.txt
scikit-learn
pandas
janome
mecab-python3
torch
transformers
fugashi
ipadic
# mecab-ipadic